refactor : QR 재발급 버튼 예약자 리스트 페이지로 이동
