+++ src/expo-admin/pages/reservations/Reservations.module.css
@@ -104,6 +104,7 @@
   font-weight: 400;
   transition: background-color 0.2s ease, color 0.2s ease;
 }
+
 .emailBtn:hover {
   background-color: #d9d9d9;
   color: #1e2a38;
@@ -116,11 +117,23 @@
   font-weight: 400;
   transition: background-color 0.2s ease, color 0.2s ease;
 }
+
 .excelBtn:hover {
   background-color: #d9d9d9;
   color: #1e2a38;
 }
 
+.qrBtn {
+  background-color: #ececec;
+  color: #1e2a38;
+  font-size: 12px;
+  font-weight: 400;
+}
+
+.qrBtn:hover {
+  background-color: #d9d9d9;
+  color: #1e2a38;
+}
 
 /* 아이콘 */
 .icon {
